Hotel_Management_App/requirements.txt
google-generativeai
python-dotenv
pandas
numpy
beautifulsoup4
chromadb
uvicorn 
fastapi 
SQLAlchemy
PyMySQL
httpx
cryptography
requests
streamlit
python-dateutil
uuid
asyncio

